Print all log messages on stderr
That's a lot more straightforward to users, as other tools do the same.

"Special messages" like the mount dir (on --appimage-mount) etc. are
still printed to stdout for easier piping and being able to
differentiate between log/error messages and values the users might
want to use for automation.

CC #923.